requirements.txt
aiohttp
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
